pyproject.toml
[tool.poetry]
name = "trading-suite"
version = "0.1.0"
description = "A modular, research-oriented end-to-end trading suite for U.S. equities using LLM agents, Deep RL, HRM, and True Confidence uncertainty module."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
yfinance = "^0.2.32"
pandas = "^2.1.3"
duckdb = "^0.9.2"
pyarrow = "^14.0.1"
great-expectations = "^0.17.10"
transformers = "^4.35.2"
torch = "^2.1.1"
mlflow = "^2.9.0"
prefect = "^2.14.9"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
alpaca-py = "^0.10.7"
backtrader = "^1.9.76.123"
stable-baselines3 = "^2.2.1"
streamlit = "^1.29.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
ruff = "^0.1.6"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
